--- a/staging/src/k8s.io/component-base/logs/json/register/register.go
+++ b/staging/src/k8s.io/component-base/logs/json/register/register.go
@@ -23,5 +23,5 @@ import (
 
 func init() {
 	// JSON format is optional klog format
-	logs.LogRegistry.Register(logs.JSONLogFormat, json.JSONLogger)
+	logs.LogRegistry.Register(logs.JSONLogFormat, &json.JSONLogger)
 }