--- a/staging/src/k8s.io/component-base/logs/registry.go
+++ b/staging/src/k8s.io/component-base/logs/registry.go
@@ -25,20 +25,21 @@ import (
 
 // LogFormatRegistry store klog format registry
 type LogFormatRegistry struct {
-	registry map[string]logr.Logger
+	registry map[string]*logr.Logger
 	frozen   bool
 }
 
 // NewLogFormatRegistry return new init LogFormatRegistry struct
 func NewLogFormatRegistry() *LogFormatRegistry {
 	return &LogFormatRegistry{
-		registry: make(map[string]logr.Logger),
+		registry: make(map[string]*logr.Logger),
 		frozen:   false,
 	}
 }
 
-// Register new log format registry to global logRegistry
-func (lfr *LogFormatRegistry) Register(name string, logger logr.Logger) error {
+// Register new log format registry to global logRegistry.
+// nil is valid and selects the default klog output.
+func (lfr *LogFormatRegistry) Register(name string, logger *logr.Logger) error {
 	if lfr.frozen {
 		return fmt.Errorf("log format is frozen, unable to register log format")
 	}
@@ -50,7 +51,7 @@ func (lfr *LogFormatRegistry) Register(name string, logger logr.Logger) error {
 }
 
 // Get specified log format logger
-func (lfr *LogFormatRegistry) Get(name string) (logr.Logger, error) {
+func (lfr *LogFormatRegistry) Get(name string) (*logr.Logger, error) {
 	re, ok := lfr.registry[name]
 	if !ok {
 		return nil, fmt.Errorf("log format: %s does not exists", name)
@@ -59,7 +60,7 @@ func (lfr *LogFormatRegistry) Get(name string) (logr.Logger, error) {
 }
 
 // Set specified log format logger
-func (lfr *LogFormatRegistry) Set(name string, logger logr.Logger) error {
+func (lfr *LogFormatRegistry) Set(name string, logger *logr.Logger) error {
 	if lfr.frozen {
 		return fmt.Errorf("log format is frozen, unable to set log format")
 	}