#### What type of PR is this?

/kind cleanup
/kind api-change

#### What this PR does / why we need it:


This replaces the experimental logr v0.4 with the stable v1.1.0
release. This is a breaking API change for some users because:
  - Comparing logr.Logger against nil is not possible anymore:
    it's now a struct instead of an interface. Code which
    allows a nil logger must switch to *logr.Logger as type.
  - Logger implementations must be updated in lockstep.
    
Instead of updating the forked zapr code in json.go, directly using
the original go-logr/zapr is simpler and avoids duplication of effort.
    
#### Which issue(s) this PR fixes:
Fixes #104097

#### Special notes for your reviewer:

The updated zapr supports logging of numeric verbosity. Error messages
don't have a verbosity (= always get logged), so "v" is not getting
added to them anymore.
    
Source code logging for panic messages got fixed so that it references
the code with the invalid log call, not the json.go implementation.
    
Finally, zapr includes additional information in its panic
messages ("zap field", "ignored key", "invalid key").

#### Does this PR introduce a user-facing change?
```release-note
Users of LogFormatRegistry in component-base must update their code to use the logr v1.0.0 API. The JSON log output now uses the format from go-logr/zapr (no `v` field for error messages, additional information for invalid calls) and has some fixes (correct source code location for warnings about invalid log calls).
```