@pohly 2021-08-03
This is WIP because klog doesn't have a formal release yet and because there are known test failures in staging/src/k8s.io/component-base/logs/json.