@pohly 2021-08-03
This is failing because it uses `Logger{}` to denote an uninitialized logger instead of the former nil, but that then cause nil pointer access because klog does not (and cannot!) detect that such a struct is unusable.