@thockin 2021-08-03
Can you use `logr.Discard()` instead, if you need a valid Logger?

On Tue, Aug 3, 2021 at 10:33 AM Patrick Ohly ***@***.***> wrote:
>
> This is failing because it uses Logger{} to denote an uninitialized logger instead of the former nil, but that then cause nil pointer access because klog does not (and cannot!) detect that such a struct is unusable.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub, or unsubscribe.
> Triage notifications on the go with GitHub Mobile for iOS or Android.