@pohly 2021-08-04
> Can you use `logr.Discard()` instead, if you need a valid Logger

No, klog needs to distinguish between "no logger" (= fall back to its own code) and "have logger" (= call logger). The discard logger is a valid logger and should be called if set.

On Slack we agreed to change code which needs to make that distinction to `*logr.Logger`. This is already done in klog internally. In k/k, the log format registry needs to be updated to that type.