@pohly 2021-08-04
At this point the only technical problem is the missing "v" field in the JSON output when using upstream zapr. This breaks the `k8s.io/kubernetes/vendor/k8s.io/component-base/logs/json` unit test. Discussion around that is in https://github.com/kubernetes/kubernetes/issues/104097