@thockin 2021-08-04
Reminder: I often miss @ mentions - the best way to catch my eye is to
assign to me or ping me issues directly :(

On Wed, Aug 4, 2021 at 7:47 AM Patrick Ohly ***@***.***> wrote:
>
> At this point the only technical problem is the missing "v" field in the JSON output when using upstream zapr. This breaks the k8s.io/kubernetes/vendor/k8s.io/component-base/logs/json unit test. Discussion around that is in #104097
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub, or unsubscribe.
> Triage notifications on the go with GitHub Mobile for iOS or Android.