@serathius 2021-09-02
This looks great! Awesome job!
Aside of adding entry to changelog about removal of "v" in error logs LGTM