@pohly 2021-09-02
> This PR may require API review.

The "api-change" label probably was a bit too conservative: while some Go APIs change, that's not tracked via that label. The JSON log format changes slightly, but that is still alpha anyway.