@pohly 2021-09-02
@kubernetes/dep-approvers: @serathius finished the technical review, now we need approval for the modified dependencies (external zapr instead of internal fork, updated dependencies).