@pohly 2021-09-09
Just noticed some left-over comment and a formatting issue, force-pushed.