@liggitt 2021-09-09
are there any benchmarks/performance tests on the structured logging to demonstrate this replumbing didn't noticeably affect resource usage?