@pohly 2021-09-09
> are there any benchmarks/performance tests on the structured logging to demonstrate this replumbing didn't noticeably affect resource usage?

I ran the `staging/src/k8s.io/component-base/logs/json` benchmarks without and with these changes and got:
```
$ benchstat /tmp/logr04 /tmp/logr10
name               old time/op    new time/op    delta
InfoLoggerInfo-36    1.84µs ± 8%    1.93µs ± 8%     ~     (p=0.246 n=5+5)
ZapLoggerError-36    2.89µs ± 2%    2.87µs ± 4%     ~     (p=0.841 n=5+5)
ZapLoggerV-36        2.23µs ± 2%    2.14µs ± 3%   -3.80%  (p=0.016 n=5+5)

name               old alloc/op   new alloc/op   delta
InfoLoggerInfo-36    1.43kB ± 0%    1.43kB ± 0%     ~     (p=0.111 n=5+5)
ZapLoggerError-36    1.62kB ± 0%    1.56kB ± 0%   -3.93%  (p=0.008 n=5+5)
ZapLoggerV-36        1.45kB ± 0%    1.43kB ± 0%   -1.21%  (p=0.000 n=5+4)

name               old allocs/op  new allocs/op  delta
InfoLoggerInfo-36      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
ZapLoggerError-36      10.0 ± 0%      10.0 ± 0%     ~     (all equal)
ZapLoggerV-36          8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
```

So performance gets slightly better.